Registration, login, and cache security settings
Security settings during registration, including password, number of characters in username and password, confirmation or not of email and phone number, etc. (of course, these settings are commented out so that the security level is not lowered); cache settings and user cache renewal and redirection to the login page if the user does not register and redirection to an error screen for a user who does not have access to a page.